+++ chaospy/distributions/approximation.py
@@ -75,6 +75,8 @@ def approximate_inverse(
     else:
         xlower = numpy.broadcast_to(bounds[0], qloc.shape)
         xupper = numpy.broadcast_to(bounds[1], qloc.shape)
+        _lower, distribution._lower = distribution._lower, lambda **kws: xlower
+        _upper, distribution._upper = distribution._upper, lambda **kws: xupper
 
     xloc = xlower+qloc*(xlower+xupper) if xloc0 is None else xloc0
     uloc = numpy.zeros(qloc.shape)
@@ -82,20 +84,21 @@ def approximate_inverse(
     uupper = 1-qloc
     indices = numpy.ones(qloc.shape[-1], dtype=bool)
 
-    if parameters is None:
-        parameters = distribution.get_parameters(idx, cache, assert_numerical=True)
-    else:
+    cache_copy = cache.copy()
+    parameters_copy = distribution._parameters.copy()
+    if parameters is not None:
+
         assert not any([isinstance(value, chaospy.Distribution)
                         for value in parameters.values()])
-        for name, param in parameters.items():
-            if isinstance(distribution._parameters[name], chaospy.Distribution):
-                cache[(idx, distribution._parameters[name])] = param
+        distribution._parameters.update(parameters)
 
     for idx_ in range(2*iterations):
 
+        cache.clear()
+        cache.update(cache_copy)
         # evaluate function:
         uloc = numpy.where(
-            indices, distribution._cdf(xloc, **parameters)-qloc, uloc)
+            indices, distribution._get_fwd(xloc, idx, cache)-qloc, uloc)
 
         # convergence criteria:
         indices &= numpy.abs(uloc) > tolerance
@@ -111,7 +114,14 @@ def approximate_inverse(
         # Newton increment every second iteration:
         xloc_ = numpy.inf
         if idx_ % 2 == 0:
-            derivative = distribution._pdf(xloc, **parameters)
+            cache.clear()
+            cache.update(cache_copy)
+            try:
+                derivative = distribution._get_pdf(xloc, idx, cache)
+            except chaospy.UnsupportedFeature:
+                cache.clear()
+                cache.update(cache_copy)
+                derivative = approximate_density(distribution, idx, xloc, cache)
             derivative = numpy.where(derivative, derivative, 1)
             xloc_ = xloc-uloc/derivative
 
@@ -129,6 +139,13 @@ def approximate_inverse(
         # print("Too many iterations required to estimate inverse.")
         # print("%d out of %d did not converge." % (numpy.sum(indices), len(indices)))
 
+    cache.clear()
+    cache.update(cache_copy)
+    distribution._parameters.clear()
+    distribution._parameters.update(parameters_copy)
+    if bounds is not None:
+        distribution._lower = _lower
+        distribution._upper = _upper
     logger.debug("%s: ppf approx used %d steps", distribution, idx_/2)
     # print("%s: ppf approx used %d steps" % (distribution, idx_/2))
     return xloc
@@ -140,7 +157,7 @@ def approximate_moment(
         distribution,
         k_loc,
         order=None,
-        rule="fejer",
+        rule="clenshaw_curtis",
         **kwargs
 ):
     """
@@ -170,7 +187,7 @@ def approximate_moment(
 
     """
     if order is None:
-        order = int(1000./numpy.log2(len(distribution)+1))
+        order = int(1e5**(1./len(distribution)))
     assert isinstance(order, int)
     assert isinstance(distribution, chaospy.Distribution)
     k_loc = tuple(numpy.asarray(k_loc).tolist())
@@ -196,7 +213,7 @@ def approximate_density(
         idx,
         xloc,
         cache=None,
-        tolerance=1e-7
+        step_size=1e-7
 ):
     """
     Approximate the probability density function.
@@ -211,8 +228,9 @@ def approximate_density(
         cache (Optional[Dict[Distribution, Tuple[numpy.ndarray, numpy.ndarray]]]):
             Current state in the evaluation graph. If omitted, assume that
             evaluations should be done from scratch.
-        tolerance (float):
-            Acceptable error level for the approximations
+        step_size (float):
+            The relative step size between two points used to calculate the
+            derivative.
 
     Returns:
         numpy.ndarray: Local probability density function with
@@ -228,17 +246,26 @@ def approximate_density(
         array([0.0242, 0.0399, 0.0242])
 
     """
-    cache = cache or {}
+    cache = {} if cache is None else cache
+    assert (idx, distribution) not in cache
     xloc = numpy.asfarray(xloc)
     assert xloc.ndim == 1
     lower, upper = numpy.min(xloc), numpy.max(xloc)
     middle = .5*(lower+upper)
-    tolerance = (numpy.where(xloc < middle, tolerance, -tolerance)*
+    step_size = (numpy.where(xloc < middle, step_size, -step_size)*
                  numpy.clip(numpy.abs(xloc), 1, None))
 
-    floc1 = distribution._get_fwd(xloc, idx, cache=cache.copy())
-    floc2 = distribution._get_fwd(xloc+tolerance, idx, cache=cache.copy())
-    floc = numpy.abs((floc2-floc1)/tolerance)
+    cache1 = cache.copy()
+    floc1 = distribution._get_fwd(xloc, idx, cache=cache1)
+    cache2 = cache.copy()
+    floc2 = distribution._get_fwd(xloc+step_size, idx, cache=cache2)
+    floc = numpy.abs((floc2-floc1)/step_size)
+
+    # weave a history of pdf from two cdf streams
+    for key in set(cache1).difference(cache):
+        cache[key] = cache1[key][0], ((cache2[key][1]-cache1[key][1])/
+                                      (cache2[key][0]-cache1[key][0]))
+
 
     assert floc.shape == xloc.shape
     return floc