+++ chaospy/distributions/baseclass/shift_scale.py
@@ -48,7 +48,9 @@ class ShiftScaleDistribution(Distribution):
             distribution=self,
             parameters=dict(shift=shift, scale=scale),
             rotation=rotation,
+            is_operator=True,
             length=length,
+            extra_parameters=dict(dist=dist),
         )
         super(ShiftScaleDistribution, self).__init__(
             parameters=parameters,
@@ -79,36 +81,46 @@ class ShiftScaleDistribution(Distribution):
         assert not assert_numerical or not (isinstance(shift, Distribution) or
                                             isinstance(scale, Distribution))
 
-        return dict(dist=self._dist, shift=shift, scale=scale,
-                    parameters=self._dist.get_parameters(
-                        idx, cache, assert_numerical=assert_numerical))
+        return dict(idx=idx, dist=self._dist, shift=shift, scale=scale, cache=cache)
 
-    def _ppf(self, qloc, dist, shift, scale, parameters):
-        return dist._ppf(qloc, **parameters)*scale+shift
+    def _ppf(self, qloc, idx, dist, shift, scale, cache):
+        return dist._get_inv(qloc, idx, cache=cache)*scale+shift
 
-    def _cdf(self, xloc, dist, shift, scale, parameters):
-        return dist._cdf((xloc-shift)/scale, **parameters)
+    def _cdf(self, xloc, idx, dist, shift, scale, cache):
+        return dist._get_fwd((xloc-shift)/scale, idx, cache=cache)
 
-    def _pdf(self, xloc, dist, shift, scale, parameters):
-        return dist._pdf((xloc-shift)/scale, **parameters)/scale
+    def _pdf(self, xloc, idx, dist, shift, scale, cache):
+        return dist._get_pdf((xloc-shift)/scale, idx, cache=cache)/scale
 
-    def _mom(self, kloc, dist, shift, scale, parameters):
-        del parameters
+    def get_mom_parameters(self):
+        parameters = self.get_parameters(
+            idx=None, cache={}, assert_numerical=False)
+        del parameters["idx"]
+        del parameters["cache"]
+        return parameters
+
+    def _mom(self, kloc, dist, shift, scale):
         poly = numpoly.variable(len(self))
         poly = numpoly.sum(scale*poly, axis=-1)+shift
         poly = numpoly.set_dimensions(numpoly.prod(poly**kloc), len(self))
-        out = sum(dist._get_mom(key)*coeff
-                  for key, coeff in zip(poly.exponents, poly.coefficients))
+        out = sum([dist._get_mom(key)*coeff
+                   for key, coeff in zip(poly.exponents, poly.coefficients)])
         return out
 
-    def _ttr(self, kloc, dist, shift, scale, parameters):
-        coeff0, coeff1 = dist._ttr(kloc, **parameters)
+    def get_ttr_parameters(self, idx):
+        parameters = self.get_parameters(
+            idx=idx, cache={}, assert_numerical=False)
+        del parameters["cache"]
+        return parameters
+
+    def _ttr(self, kloc, idx, dist, shift, scale):
+        coeff0, coeff1 = dist._get_ttr(kloc, idx)
         coeff0 = coeff0*scale+shift
         coeff1 = coeff1*scale*scale
         return coeff0, coeff1
 
-    def _lower(self, dist, shift, scale, parameters):
-        return dist._lower(**parameters)*scale+shift
+    def _lower(self, idx, dist, shift, scale, cache):
+        return dist._get_lower(idx, cache=cache)*scale+shift
 
-    def _upper(self, dist, shift, scale, parameters):
-        return dist._upper(**parameters)*scale+shift
+    def _upper(self, idx, dist, shift, scale, cache):
+        return dist._get_upper(idx, cache=cache)*scale+shift