+++ chaospy/distributions/baseclass/simple.py
@@ -1,4 +1,5 @@
 """Distribution for the core probability distribution."""
+import itertools
 import numpy
 import chaospy
 
@@ -75,8 +76,32 @@ class SimpleDistribution(Distribution):
         for key, value in parameters.items():
             if isinstance(value, Distribution):
                 value = value._get_cache(idx, cache, get=0)
-                assert not assert_numerical or not isinstance(value, Distribution)
+                if isinstance(value, Distribution) and assert_numerical:
+                    raise chaospy.UnsupportedFeature(
+                        "operation not supported for %s with dependencies" % self)
                 parameters[key] = value
+        return parameters
+
+    def get_upper_parameters(self, idx, cache):
+        parameters = self.get_parameters(idx=idx, cache=cache, assert_numerical=False)
+        for key, value in parameters.items():
+            if isinstance(value, Distribution):
+                parameters[key] = value._get_upper(idx, cache)
+        return parameters
+
+    def get_lower_parameters(self, idx, cache):
+        parameters = self.get_parameters(idx=idx, cache=cache, assert_numerical=False)
+        for key, value in parameters.items():
+            if isinstance(value, Distribution):
+                parameters[key] = value._get_lower(idx, cache)
+        return parameters
+
+    def get_mom_parameters(self):
+        parameters = self.get_parameters(
+            idx=None, cache={}, assert_numerical=False)
+        if any([isinstance(value, Distribution) for value in parameters.values()]):
+            raise chaospy.UnsupportedFeature(
+                "operation not supported for %s with dependencies" % self)
         return parameters
 
     def _mom(self, kloc, **kwargs):