+++ chaospy/distributions/baseclass/user.py
@@ -1,4 +1,5 @@
 """Distribution with user-provided methods."""
+import numpy
 import chaospy
 
 from .simple import SimpleDistribution
@@ -73,11 +74,26 @@ class UserDistribution(SimpleDistribution):
         """
         self._cdf = cdf
         repr_args = [str(cdf)]
-        for name in ["pdf", "lower", "upper", "ppf", "mom", "ttr"]:
-            value = locals()[name]
-            if value is not None:
-                repr_args.append("%s=%s" % (name, value))
-                setattr(self, "_"+name, value)
+        if ppf is None and (lower is None or upper is None):
+            raise chaospy.UnsupportedFeature("either ppf or lower+upper should be provided.")
+        if pdf is not None:
+            repr_args.append("pdf=%s" % pdf)
+            self._pdf = pdf
+        if lower is not None:
+            repr_args.append("lower=%s" % lower)
+            self._lower = lower
+        if upper is not None:
+            repr_args.append("upper=%s" % upper)
+            self._upper = upper
+        if ppf is not None:
+            repr_args.append("ppf=%s" % ppf)
+            self._ppf = ppf
+        if mom is not None:
+            repr_args.append("mom=%s" % mom)
+            self._mom = mom
+        if ttr is not None:
+            repr_args.append("ttr=%s" % ttr)
+            self._ttr = ttr
 
         parameters = parameters if parameters else {}
         if parameters:
@@ -87,3 +103,15 @@ class UserDistribution(SimpleDistribution):
 
         super(UserDistribution, self).__init__(
             repr_args=repr_args, parameters=parameters)
+
+    def _lower(self, **parameters):
+        x_loc = 0.
+        for param in parameters.values():
+            x_loc, _ = numpy.broadcast_arrays(x_loc, param)
+        return self._ppf(x_loc, **parameters)
+
+    def _upper(self, **parameters):
+        x_loc = 1.
+        for param in parameters.values():
+            x_loc, _ = numpy.broadcast_arrays(x_loc, param)
+        return self._ppf(x_loc, **parameters)