+++ chaospy/distributions/operators/addition.py
@@ -65,9 +65,9 @@ Inverse transformations::
 Raw moments::
 
     >>> joint1.mom([(0, 1, 1), (1, 0, 1)]).round(4)
-    array([ 6.    ,  2.5   , 15.0833])
+    array([ 6.    ,  2.5   , 15.0834])
     >>> joint2.mom([(0, 1, 1), (1, 0, 1)]).round(4)
-    array([ 6.    ,  3.5   , 21.0833])
+    array([ 6.    ,  3.5   , 21.0834])
 """
 from __future__ import division
 from scipy.special import comb