+++ chaospy/distributions/operators/joint.py
@@ -18,7 +18,6 @@ class J(Distribution):
     """
 
     def __init__(self, *args, **kwargs):
-        dists = []
         repr_args = []
         owners = {}
         index = []
@@ -34,8 +33,8 @@ class J(Distribution):
                 idx += len(arg)
             elif len(arg) > 1:
                 repr_args.append(arg)
-                owners.update({idx2: (idx1, arg)
-                               for idx1, idx2 in enumerate(range(idx, idx+len(arg)))})
+                owners.update({idx2: (idx1, arg) for idx1, idx2 in enumerate(
+                    range(len(index), len(index)+len(arg)))})
                 index += [idx]*len(arg)
                 idx += 1
             else:
@@ -147,7 +146,7 @@ class J(Distribution):
             >>> d0 = chaospy.Uniform()
             >>> dist = chaospy.J(d0, d0+chaospy.Uniform())
             >>> dist.mom([1, 1]).round(4)
-            0.5833
+            0.5834
             >>> dist = chaospy.J(chaospy.Uniform(), chaospy.Normal())
             >>> dist.mom([[0, 0, 1], [0, 1, 1]]).round(4)
             array([1., 0., 0.])