+++ tests/distributions/operators/test_addition.py
@@ -149,10 +149,9 @@ def test_dist_addition_recurrence():
 
 def test_dist_addition_wrappers():
     dists = [
-        chaospy.Normal([2, 2, 3]),  # ShiftScale
-        chaospy.Uniform(1, [3, 3, 5]),  # LowerUpper
+        chaospy.J(chaospy.Normal(2), chaospy.Normal(2), chaospy.Normal(3)),  # ShiftScale
+        chaospy.J(chaospy.Uniform(1, 3), chaospy.Uniform(1, 3), chaospy.Uniform(1, 5)),  # LowerUpper
         chaospy.MvNormal([2, 2, 3], numpy.eye(3)),  # MeanCovariance
-        chaospy.J(chaospy.Uniform(1, 3), chaospy.Uniform(1, [3, 5])),  # Joint
     ]
     for dist in dists:
         joint = chaospy.Add([1, 1, 3], dist)